requirements.txt
colorama==0.4.6
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
tk==0.1.0
tomli==2.0.1
